# payments: allocate payment sequences in blocks