--- a/docs/release-notes/release-notes-0.14.0.md
+++ b/docs/release-notes/release-notes-0.14.0.md
@@ -24,6 +24,9 @@ the release notes folder that at leasts links to PR being added.
   code](https://github.com/lightningnetwork/lnd/pull/5547) when using etcd
   backend.
 
+[Optimized payment sequence generation](https://github.com/lightningnetwork/lnd/pull/5514/)
+to make LNDs payment throughput (and latency) with better when using etcd.
+
 # Contributors (Alphabetical Order)
 * ErikEk
 * Zero-1729