@bhandras 2021-07-21
> I also like the idea and the performance gain!
> 
> A few typos and nits, plus I'm not so sure about the idea of skipping integers in a sequence and its implications. But the performance gain is nice, from my local machine,
> 
> ```
> master
> === RUN   TestLightningNetworkDaemon/37-of-82/btcd/async_payments_benchmark
>     test_harness.go:120:    Benchmark info: Elapsed time:  29.371069952s
>     test_harness.go:120:    Benchmark info: TPS:  16.444753316421505
> 
> this fix
> === RUN   TestLightningNetworkDaemon/37-of-82/btcd/async_payments_benchmark
>     test_harness.go:120:    Benchmark info: Elapsed time:  27.936773974s
>     test_harness.go:120:    Benchmark info: TPS:  17.2890399030867
> ```
> 
> And got a question about the plan on other sequencers? Do we plan to optimize them too? For instance, when initializing payments we use this `nextPaymentSequence`. For HTLC attempts, there's the `persistentSequencer`, which uses the bucket `next-payment-id-key` to manage the IDs. Since a payment might have multiple HTLC attempts, we might achieve even more performance gain there.

Interesting I wasn't aware of the `persistentSequencer` but it looks very similar. Can I ask if you tested with `dbbackend=etcd`? I just re-ran the test and for me the performance gain is substantial (same as in the PR descripton).